+++ src/index.css
@@ -3,3 +3,13 @@ body{
     padding: 0;
     border: 0;
 }
+a{
+    color:black;
+}
+a:hover{
+    color:black;
+    text-decoration: none;
+}
+.selectedTab{
+    background-color: #e9ecef;
+}
